pulling out the static config provider too
